chore(Rules & AboutUs): Modify groups descriptions and rules
The /aboutus page now contains the new groups "Recursos" and "Mesa de Ayuda" as part of the groups descriptions, reflecting the new structure the community has.

The /rules page shows a detailed description of the rules for these new groups as well.

Some profile pictures were deleted and updated. This reflects the current structure of the administration group.